Store users
### What's done:
* Code style
