+++ save-backend/src/main/kotlin/org/cqfn/save/backend/configs/WebSecurityConfig.kt
@@ -5,16 +5,13 @@
 package org.cqfn.save.backend.configs
 
 import org.cqfn.save.backend.utils.ConvertingAuthenticationManager
-import org.cqfn.save.backend.utils.GithubOauthTokenIntrospector
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Profile
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity
 import org.springframework.security.config.web.server.ServerHttpSecurity
 import org.springframework.security.crypto.factory.PasswordEncoderFactories
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.security.web.server.SecurityWebFilterChain
-import reactor.core.publisher.Mono
 
 @EnableWebFluxSecurity
 @Profile("secure")
